Rewrote Titan-Hadoop incremental loading docs
This feature has been restored as of
bf39926f3f10ad8aacb16971d545c8595894e88c, but without documentation.
Incremental loading is now documented in roughly the same place as it
used to be in 0.4.4 (under the Titan I/O format docs), but the prose
and code examples have been rewritten from scratch.