Tweak default cassandrathrift connection pool settings
This commit disables the maxTotal parameter on the connection pool.
The connection pool size is now regulated through the maxActive and
maxIdle parameters on the pool.  Setting maxTotal appears to cause
commons-pool to repeatedly just barely exceed the limit by one
connection, then close 15% of the total pool size before repeating the
process.  The overall number of connections stays at or just one over
the configured limit, which is good, but routinely closing and
reopening connections can lead to a tens of thousands of TIME_WAIT
sockets in test scenarios with just one client and one server on the
same physical host, which is bad.

Keeping maxTotal off (negative) and using the active and idle limits
instead seems to avoid this behavior.  I've seen this before when
testing Titan-Hadoop and worked around it there by overridding the
defaults just for the affected Titan-Hadoop testsd.  Now I'm seeing it
again in ThriftGraphTest's getSlice stress tests, and I'm changing the
defaults instead of just working around it in a specific test config.