Copy commented Titan-Hadoop config tips
The stock config file conf/hadoop/titan-graphson.properties had a note
about overridding Hadoop configuration settings along with some
commented suggested Hadoop config key-value pairs to consider.  This
commit copies that chunk of comments to the other stock Titan-Hadoop
config files.

This should be templated one way or another so that we don't have to
maintain a bunch of copies of the same chunk of text.