Suppress shutdown hook exception
The code that stats an embedded Cassandra daemon for test purposes
also installs a shutdown hook to kill the daemon.  The hook calls a
method which can also be called from a separate stop method by an
ordinary thread outside the JVM shutdown hook sequence.  This method
attempts to uninstall the shutdown hook once it's done (useful on the
stop path).  However, uninstalling a shutdown hook when the JVM is
trying to exit (the hook path) results in an IllegalStateException.
It's scary looking but not harmful.  This commit makes the method
tolerate and log ISE.