Fix convenience getters in FaunusSchemaManager
FSM defines a general getRelation method, plus convenience methods for
specific relation subtypes.  FSM is just a pair of dumb
ConcurrentHashMaps with no awareness of Titan schema.  FSM.getRelation
delegates to a helper class that knows about Titan's ManagementSystem
API when a retrieval on one of the maps misses.  However, the subtype
helper getters did not delegate.  This led to a confusing NPE in
TitanHadoopGraph.readHadoopVertex's call to
vertex.getTypeManager().getPropertyKey(name); the type was defined in
FSM's helper class, but FSM was ignoring it.  I think this was
unintentional, since the contains methods on FSM also consult the
ManagementSystem-aware helper.