Change context type in incr load method signature
This is probably going to be a short-lived interface with TP3 around
the corner, but the incremental loading script should probably use the
more general TaskInputOutputContext interface as opposed to
Mapper.Context or Reducer.Context, in case we want to call it from a
Reducer at some point.  TaskIOContext is the most general interface
that still allows getCounter under both Hadoop 1 and 2.