Removing/tidying some old TP 2.x test code
This commit removes:

* FakeVertex and associated Frames-based test in
  TitanGraphSerialSpeedTest

* GraphGenerator.  This was really just a thin wrapper around the
  power law graph generator in an old alpha version of Furnace.
  TitanGraphSerialSpeedTest continues to use a graphml file produced
  by GraphGenerator a while ago (titan-test/data/v10k.graphml.gz);
  only the generator that produced it is deleted in this commit.

* The alpha Furnace code in titan-test (used only by GraphGenerator)

I shortened "SerialSpeedTest" to just "SpeedTest" in the classes where
it appears.  I also renamed the ludicrously vague "Schema" class used
by the (Serial)SpeedTest to SerialSpeedTestSchema and moved it into
the same package and folder that uses it.  It's now a Groovy class.

I temporarily commented the body and imports of TitanIndexTest so that
I could get titan-test to run its Grooovy compiler.  Then I updated a
couple of imports (e.g. BP 2.5 Vertex is now TP3 structure.Vertex).
The (Serial)SpeedTest groovy code compiles now:

[INFO] --- gmaven-plugin:1.5:compile (groovy-compile) @ titan-test ---
[INFO] Compiled 18 Groovy classes

It probably has runtime failures, but at it shouldn't cause compile
errors or get in the way of updating & compiling other code in
titan-test to TP3.

TitanIndexTest is not modified in this commit.  I only commented it in
my working copy so that Maven would stop choking on it and run the
gmaven-plugin compilation step.