Restoring HBase support in Titan-Hadoop
This commit adds an HBase VertexInputFormat similar to the one
introduced for Cassandra in 5a9ed8ef8590190deb8085c33984906bc4e7190b.
Both formats (and titan-hadoop in general) are tightly coupled to
TinkerGraph and loosely coupled to GiraphComputeVertex.

This titan-hadaoop change only been tested on HBase 0.98.3-hadoop1.

This commit also modifies TitanGremlinPlugin to pull in
hbase-{client,common,server} version 0.98.3-hadoop1, which is a
complete hack for testing and not the right way to manage that
dependency.  hbase-server is needed only for HBase's TableInputFormat.