Adding six hour fork timeout to Surefire config
This is mostly for our CI platform, which is capable of detecting hung
builds based on the amount of time passed since their last line of
output, but which nonsensically reliant on a nonstandard plugin to
kill the jobs once flagged as hung.  And the plugin only works on a
manual bring-your-own-hardware install, not the CI-as-a-service cloud
option.

This is specifically for a strange deadlock in
TransactionalGraphTestSuite (often line 484 when it deadlocks and BDB
detects the cycle) that shows up in CI but not on my workstation.  The
root cause wasn't completely obvious at first read.
(cherry picked from commit 736d050e5e924fbde3a59e83520e55c02fbda088)