Make STG's shutdown hook avoid removing itself
This is for #842 on 0.9.0-SNAPSHOT.  The shutdown hook now carries out
all shutdown steps except hook removal.  The shutdown method on STG
does everything the hook does plus hook removal.

This commit is a rewrite of 1c37382e35f9d0901ed67d26aace6c539d049c82.
The fundamental approach is identical, but the code has changed enough
between titan05 and titan09 that cherry-picking it would be slower
than just editing it in titan09.