Updating some Gremlin snippets in the docs for 0.5
The most important change in this commit is replacement of some
Gremlin `(iterable of vertices).type` calls with `(iterable of
vertices)*.getVertexLabel()`.  I also ran through the getting-started
steps in the intro again on 0.5 and copy-pasted the output as it
appears today: type=foo no longer appears in `map` output, calling
`map` on a single vertex lists each property on a separate line (used
to be all on one line enclosed by curly braces apparently), and edge
toStrings are formatted slightly differently.  The rest of the changes
should just account for different vertex IDs than used to be assigned
in the getting-started scenario.