Partitioned vertex filtering for TP3 input formats
Added an option to the HBase/Cassandra
InputFormat<NullWritable,VertexWritable> implementations built for TP3
that controls the treatment of Titan's partitioned vertices.  By
default, the presence of a partitioned vertex in the input causes an
exception.  Setting titanmr.input.filter-partitioned-vertices=true
instead drops partitioned vertices and any edges incident on
partitioned vertices, logging each at debug as it is dropped.

This commit also renames several config keys in
TitanHadoopConfiguration and adds a test of the new partitioned vertex
filter setting.  The test code is a bit clumsy since I've tried to
avoid requiring Giraph for the titan-hadoop tests -- this could
probably be streamlined once TP3's MapReduceGraphComputer is
implemented.