Depend on Objenesis
This was excluded from our dependency on older versions of Kryo, since
Kryo used to include a shaded copy of Objenesis.  As of 2.24.0, Kryo
doesn't include a shaded copy and needs an Objenesis dependency to
work (judging from the list of classfiles in the 2.24.0 jar on
Central).

Thanks @superbeeny

For #828