Throw new ExpiredLockException when applicable
This commit modifies ConsistentKeyLocker and its superclass to
implement the new behavior expected in
6ea5ebf8d0b13c025d08a4734037fb725d0e089c.  Previously, a lock that
expired while a transaction was holding it would throw a cryptic
TemporaryLockingException with the message "No lock columns found for
KeyColumn [k=foo, c=bar]".  This is not a helpful exception in most
cases; the log for ConsistentKeyLocker emits a WARN about the expired
lock shortly before this is thrown, but it would be better to have a
specific exception type and message to make it clearer about what's
happening in this case.

Now we have ExpiredLockException.  This is a a subclass of
TemporaryLockingException.  The message string used with it is
something like "Expired lock on KeyColumn [k=foo, c=bar]: raw lock
timestamp 1418915175000000 MICROSECONDS is older than Duration[300000
ms]".
(cherry picked from commit 24f93bbdfff7b20a0e11bfd747813ba1a48111cb)