Add failing test with vlabel and Contain.IN conds
cherry-pick was able to apply this diff without errors, but changes in
titan09 meant that it produced compile errors even though the
diff/patch process was clean.  Required a bit of manual API fixup
around vertex labels and commits, but no changes to the substance of
the test.

This is based on #882
(cherry picked from commit 5c8161150f04f33ce40e4fc6981d72dc79f1970b)