Building titan09 + TP3 zipfiles with maven again
This commit gets the titan-dist/titan-dist-hadoop-1 module to produce
a usable gremlin-server.sh and gremlin.sh script again.  The
aurelius.titan and tinkerpop.server modules are activated by default.
There are a lot of rough edges: gremlin.sh needs a log4j config
update, the solr config files are of no use under Hadoop 1, the conf
folder is sort of a chaotic mess with a combination of ES + Cassandra
+ client + server files, etc.  It's a baseline commit.