Removing custom shorthand registration
This is a follow-on to 110e22c818314f65ab9a0d27cb4ab429abda0db6

I just changed the resource containing shorthand->class mappings to a
fully-qualified package-and-filename string.  That dramatically
reduces the odds of accidentally overwriting the file with nonsense,
but it also provides no clean way for third-party implementors to add
mappings.  I don't think centralizing this information in a file is
the way we want to approach custom shorthands anyway.  Removing the
custom shorthand feature and making the index and storage backend
mappings immutable again.

This commit does not affect the standard shorthands.  For example,
"cassandra" is still a valid shorthand.

For #929