Make ThriftGraphProvider start embedded Cass
TP3's test runner expects to be able to open a graph before running
the @BeforeClass methods on the test.  Tests that depend on
ThriftGraphProvider, including ThriftStructureTest, would fail when
run in isolation because embedded Cassandra wasn't started before the
test runner tried to open a graph, despite @BeforeClass methods on the
test.