Add some file excludes to Cassandra's SSL tests
In titan05, the tests that depended on Cassandra configured for SSL
were marked through a JUnit @Category.  This still mostly works,
except that tests @RunWith the TP3 Runner seem to ignore the @Category
list and run everything.  This leads to spurious test failures when
test code that assumes a non-SSLized Cassandra ends up talking to a
SSLized Cassandra.  As a short-term workaround, this commit adds
excludes for the test classes that use the TP3 runner under the
titan-cassandra ssl-test execution.