Fix race condition in StandardScannerExecutor
StandardScannerExecutor's dataPuller threads read binary entries from
the backend and enqueue them for consumption by processor threads.  It
was possible, prior to this commit, for the dataPullers to read from
the backend, enqueue the binary data, then trigger termination of the
processor threads before the processors ever had a chance to drain the
queues.  The processor threads now check that their queue is empty
before terminating.