Log local lock holder stacks at TRACE
When LocalLockMediator's log severity is at least TRACE, a successful
invocation of the lock method generates and stores an array of
StackTraceElement.  If a later attempt to lock fails because of a
conflict, then both stacktrace of the failed caller and the earlier
stored stacktrace of the successful caller will be logged.