Shade Kryo 2.24.0
This commit does something similar to the following two commits on
titan05:

* c736cf7dba4b72ba00c9cf8dbe2941be6e9c66a9
* e6e2dacc2b10d4936ebf1ff9d9bdb4d2498ed0d4

But it's not exactly the same.  The contents of the Kryo 2.24.0 jar
used in titan09 (relevant to this commit) are slightly different from
the contents of the 2.22 jar (relevant to above commits on the 0.5
line).  This commit is specifically for titan09/2.24.0.