Reenable and tweak PageRank Fulgora test
For #989

This PR implementation is almost an exact copy of the impl that ships
with TP3.  I've simplified the test data slightly to match the TP3
impl (removing edges with a label that was ignored under the
oldfulgora version) rather than attempting to enrich the PR impl with
additional traversal selectivity.