Addition compile error fixes
The only remaining compile errors in TitanGraphTest following this
commit should be inside testTinkerPopOptimizationStrategies and the
private helper method called only by that method (assertNumStep).  It
is possible to refactor this test to be TP3-strategy-oblivious, but
then much of its distinctive value would be lost.  Leaving it broken
for the time being.

I gave this only one close read through the git-diff before
committing.  This hasn't been tested at all.  There could easily be
some runtime errors in here that I missed.