Remove inactive @After|BeforeClass in TP3 tests
This is probably not the best way to fix it.  Not sure if TP3 has an
equivalent to @AfterClass... haven't found it yet.

This commit changes the failure mode of several HBase tests.  Instead
of failing before any test methods even run due to neglecting to start
HBase, the tests now start HBase and then fail substantial assertions.