Reworking TTL options to be like EntryMetaData
EntryMetaData is still not a totally satisfactory model for this,
since it relies on an EnumMap with values of various types, and the
client code is ultimately responsible for typecasting those values.
The point of this commit was just to unify the pattern used for entry
and store metadata like TTL, rather than having two different
patterns.