Avoid a warning about an invalid config option
GraphDatabaseConfiguration.getLocalConfiguration sets the
"gremlin.graph" property to TitanFactory's full package and classname.
MapReduceIndexManagement doesn't need this key, and it leads to a
scary warning when copying the graph's local configuration into a
Hadoop configuration object for use in an index management job.
Removing this key to avoid the warning.