TTL: method rename, condition tweak, add comments
The only substantial change in this commit is narrowing one condition
in KCVSLogManager's constructor.  Prior to this commit, the condition
assumed that StoreFeatures.hasCellTTL implied
!StoreFeatures.hasStoreTTL.  This is true for all of the storage
backends we have today, and even if it weren't true, it wouldn't
result in incorrect operation, but it's still not necessarily true in
general.  Changed condition to check for StoreFeatures.hasCellTTL &&
!StorFeatures.hasStoreTTL, which was the original intent.

The other changes are just cosmetic:

* Added some comments

* Removed a redundant precondition in TTLKCVSManager's constructor

* Renamed TTLKCVSManager's supportsStoreTTL method to supportsAnyTTL.
  This is better because the method's return value is actually cellTTL
  || storeTTL... it's confusing to name that supportsStoreTTL when it
  really means "supportsCellOrStoreTTL".