Restore @BeforeClass hack for Cassandra+TP3 tests
The TP3 test suites do not appear to have a analog to JUnit's standard
@AfterClass and @BeforeClass.
897648e14112aa16e1e10b807caceb6ee37dd233 introduced a hack to start
embedded Cassandra by overriding some ThriftGraphProvider logic that
by design should be limited to constructing a configuration.
114886e46c4e5c8a1151cd202103db0673a2c228 commented that hack, but in
discussion, it sounds like this was unintentional.

Meanwhile, I've started a thread on dev@tp to figure out whether
there's a better way to do @BeforeClass (or to do @AfterClass at all).