HBaseKCVS.RowIterator rowkeys need not be longs
RowIterator passed all rowkeys through a filtering predicate that
required them to successfully deserialize as longs.  This is not
always true: running a StandardScannerExecutor scanjob to repair a
graphindex will retrieve rowkeys whose contents are dictated by the
values of the indexed type (certainly not necessarily longs).  Some
values may be shorter than 8 bytes, and I'm indeed seeing buffer
underruns in RowIterator's filtering predicate when
HBaseGraphTest.testIndexUpdatesWithReindexAndRemove tries to repair a
graphindex.

This commit retains the condition that rowkeys be nonnull, but removes
the constraint that they be interpretable as longs.