Tweak bulk loader config handling
This is for compatibility with SparkGraphComputer.  In particular, I
seem to need the useGraphConfig method when executing on
SparkGraphComputer, even though it's not required on
GiraphGraphComputer.  I think {Spark,Giraph}GraphComputer may have
some kind of discrepancy in how they handle VP configs that bulk
loader is triggering (maybe bulk loader is doing something odd?)