Add properties-only TitanOutputFormat (Hadoop 1)
This commit adds a new Hadoop 1
OutputFormat<NullWritable,VertexWritable> that can only persist vertex
properties.  If the OF is able to instantiate a TP3 VertexProgram from
the configuration in its TaskAttemptContext, then it assumes its
running after a VP, and persists only that VP's element compute keys.
Otherwise, it tries to persist all properties.

This output format is experimental and barely tested.