Prefer HBase 0.98's hadoop 1 build over 0.94
This affects the titan-all artifact, the Hadoop-1 based titan zipfile
distribution, and the set of jars installed into $REPO_ROOT/lib when
running `mvn clean install` in a git clone of the titan repository.

The point is to avoid HBase 0.94's dependency on a very old Netty
version, which causes problems when using Spark.