Hack to make HBase 0.94 tolerate Guava 15.0
I used to get around this problem with shading for test purposes, but
7b3de6c84b55f869f087ae3a67379a1b54aad8a7 broke that strategy by
leaking code exclusive to Guava 15.0+ into core.  Now all ID
allocations fail with NoSuchMethodError on Stopwatch.createStarted
when I use the old HBase 0.94 + Guava 14 test setup.

Now that it's broken and I'm trying to run TTL tests in every module
that supports them, I need to revisit it.  HBase 0.94 refuses to start
without Guava's NullOutputStream, which was removed in 15.0.
NullOutputStream is a simple and tiny little class with no
intrapackage dependencies.  It's 39 lines including license header.
Tentatively, this seems to be all HBase 0.94 server needs added onto
Guava 15.0 to startup and server queries for test purposes.  It's
impossible to exhaustively test this, since we're not realistically
going to have full coverage through both Titan's code and all of the
HBase 0.94 code it would exercise in production.