Change TitanBlueprintsGraph.configuration
This previously returned
GraphDatabaseConfiguration.getLocalConfiguration.  This is not
necessarily identical to the configuration supplied to
TitanFactory.open if some of the options are nonlocal.  However,
org.apache.tinkerpop.gremlin.structure.GraphConstructionTest's method
shouldMaintainOriginalConfigurationObjectGivenToFactory requires that
TBT/STG.configuration return a (possibly improper) superset of the
configuration passed to the open method.

I changed GraphDatabaseConfiguration to retain the ReadConfiguration
supplied during its construction and return that through
TBT.configuration.  This is kind of a cop-out, but the test passes
now.

For #1067 part 1