Enable partitioning even on non-key-ordered stores
Prior to this commit, ID partitioning was enabled by default only on
key-ordered stores and those stores with a local partition (HBase &
Cassandra-BOP).

In discussion with @mbroecheler, it turned out that this was
unintentional: "I did want to enable partitioning on non keyordered
stores".

This commit enables unconditional ID partitioning.  It also causes the
formerly failing *GraphPartitionTest classes on C-BOP to pass.