Fix graph leak in Hadoop InputFormats
Cassandra|HBaseInputFormat created a TitanVertexDeserializer, which in
turn created a StandardTitanGraph.  However, individual RecordReaders
were responsible for closing the TVD/STG.  There are two problems with
this:

1. The first RecordReader to close will shutdown a graph potentially
   shared among many RRs

2. Client code which creates an InputFormat but no RecordReader (for
   instance, when retrieving splits) will leak an unclosed STG

Refactored this so that TVD uses in the RecordReaders are atomically
refcounted.  One TVD is still shared among potentially multiple RRs,
but now only the last RR to close it will end up calling STG.close.