Adding 60s connect timeout after starting HBase
I've never seen HBaseStorageSetup.startHBase race with the test
methods that typically follow right afterward, but I think the
potential for a race exists.  I think this may affect a CI
environment.  Introducing a loop near the end of startHBase that
attempts to establish an HConnection for up to 60 seconds after
forking the daemon before returning normally (or timing out, logging a
warning, and returning anyway).