Renamed the "g" to "graph"
`g` is pretty confusing as it is generally used to traverse aka: `graph.traversal()`. I changed the name in the configuration file to `graph` so that it would follow in the steps of the gremlin-server configuration and is more intuitive for the users. 
Anyone who needs to traverse `g` can set a script to do so. (maybe we should set one by default? `g = graph.traversal()` ?)