Handle values with dollar sign in an elastic search index
This code still appears to be a bit open to exploits, but the basic
problem is that dollar signs are passed through directly into
the groovy script to be executed. This causes groovy to fail
at executing the script because typically what comes after
the dollar sign will not correspond to a groovy variable
available in the script execution environment (nor would we
want that to work in any case). This change just adds a quick
hack to escape the dollar sign.