Initial HBase 1.0 support
Expanded the HBaseCompat shims enough to hide the ABI and API
differences across the HBase [0.94, 1.0.2] version range.
HConnection, HTable, and HBaseAdmin have been renamed or are
deprecated and in the process of being renamed.  Most of their methods
have the same signatures and functionality.  However, some other
classes that were mostly stable also underwent minor method signature
tweaks that break the ABI.  For instance, Delete.setTimestamp used to
return void.  Now it returns a Delete (this).  In testing, this
signature change caused MethodNotFoundException to be thrown when code
compiled for one signature ran against the other (even though it
ignored the return value).

titan-hbase-10/src/test/bin is a copy of 1.0.2's files with classpath
and jruby location tweaks.  For future reference, the tweaks are
stored in patch form (based on 1.0.2) inside the directory.

The only testing I've done is to check that HBaseGraphTest passes for
0.94, 0.98, and 1.0.  I ignored 0.96 since it's technically EOL.

This commit adds two fake copies of a pair of Guava classes
titan-hbase-care: Stopwatch and Closeables.  HBase depends on Guava
12, which is substantially behind titan09's current 16, let alone
Guava's latest stable 18; both of these Guava classes underwent
compatibility-breaking changes between 12 and 16.  The HBase code
compiled against Guava 12 produces runtime errors when executed
against Guava 16:

* the nullary Stopwatch constructor is no longer public (replaced by
  static methods that invoke the constructor)

* Stopwatch.elapsedMillis() was removed (replaced by
  elapesed(timeunit))

* Closeables.closeQuietly was removed (replaced by close(boolean
  swallowExceptions))

The fake copies of Stopwatch and Closeables in this commit adds the
deleted Guava functionality that HBase still depends upon (in the case
of the nullary Stopwatch constructor, it just changes the visibility
level from package to public).

To force the fake classes to take precedence, the shell scripts now
build $CLASSPATH with all jars matching the glob "titan*.jar"
(including titan-hbase) at the beginning and remaining jars (including
Guava) at the end.

For #1065