Use TP 3.0.1-SNAPSHOT
3.0.1-incubating, the release which will eventually succeed this
SNAPSHOT version, is currently undergoing a vote on the TP3 dev list
to determine whether the current release candidate becomes final.

We need some minor TP 3.0.1 changes that aren't in 3.0.0-incubating
(e.g. for #730).  Using the SNAPSHOT for the interim between now and
3.0.1-incubating's release.

This commit also adds a <repository> stanza to the top-level pom
pointing to
https://repository.apache.org/content/repositories/snapshots.