Exclude servlet-api from hadoop/hbase
The classpath used for titan-hadoop-2 tests had multiple copies of the
javax servlet API on the classpath, leading to runtime exceptions in
the tests that use SparkGraphComputer complaining that
javax.servlet.FilterRegistration's "signer information does not match
signer information of other classes", which was apparently caused by
different versions of the servlet API providing fragments of a single
package.  This is provided at runtime, and it might be the case that I
should redeclare servlet-api with <scope>runtime</scope> to make that
explicit in Maven's dependency system, but servlet-api deps are
already such a mess that I'm concerned that will just cause additional
problems somehow.  We still have multiple, possibly-conflicting
versions of it in the broader titan dependency tree, as shown when
building a zipfile or $REPO_ROOT/lib out of titan-dist.