MapReduce index jobs: support custom Hadoop conf
The most flexible and general way to support this on titan09 seems to
be to let the user supply an arbitrary o.a.hadoop.conf.Configuration
object which becomes the basis for the index management MapReduce job
that the user intends to run.  This lets the user load custom Hadoop
options as classpath resources, input streams, file paths, or as
object instances.

Options set by the MapReduce index jobs themselves (e.g. to tell the
job the index's name) still override options from the user's
o.a.hadoop.conf.Configuration.  The idea is that this feature supports
changing arbitrary options that affect the job's performance or
robustness that are outside the immediate scope of the index
management job itself (e.g. cassandra.thrift.framed.size_mb).

Fixes #1100