Avoid interrupting ID allocations on BDB
This commit introduces StoreFeatures.supportsInterruption specifically
for BerkeleyDB JE.  BDB responds with undefined behavior when a thread
is interrupted while inside a blocking IO call.  "Undefined behavior"
in this case tends to mean that all subsequent calls to the
environment (barring a manual recovery operation) appear to fail with
exceptions.

As of this commit, C* and HBase return true for supportsInterruption,
while BDB returns false.

The only place I've found so far where this matters is ID allocation.
StandardIDPool works by starting a loop-forever background task that
attempts to call getIDBlock until the first time it succeeds.
StandardIDPool then watches the clock and attempts to cancel the
background task if it has exceeded its timeout.  Cancelation sent a
thread interrupt, which in turn broke BDB JE's environment.

StandardIDPool now has a collaborative/discretionary background task
cancellation mechanism: it sets a volatile boolean field to true when
it wants the background task to shutdown, and the background task
checks it during its loop between each blocking call.  This field is
always set now when StandardIDPool wants the background task to abort.
Furthermore, StandardIDPool checks StoreFeatures.supportsInterruption
(actually, it calls IDAuthority.supportsInterruption, which wraps a
call to the same-named method on its underlying store features).  If
interruption is allowed, it calls future.cancel(true) on the task,
same as prior to this commit.  If interruption is not allowed, then
StandardIDPool adds the future to a deque.  This deque is drained and
`get` invoked on each element during StandardIDPool's close method.
This last queue-draining ensures that a clean shutdown doesn't leave
BDB in an inconsistent state because a background task was still
mid-IO-call.

Fixes #1148