Support deprecated ES sorting option
"ignore_unmapped" was deprecated in Elasticsearch 1.4.0 in favor of
the new "unmapped_type" parameter, but "ignore_unmapped" is still
useful if the Elasticsearch server version is older than 1.4.0.  Those
older server versions will fail requests with an error if the requests
attempt to use the new "unmapped_type" option.  Adding support for the
older, deprecated ES sort option.  It must be manually enabled in the
configuration; Titan does *not* attempt to detect the Elasticsearch
node versions using a NodeInfoRequest.

By default, Titan still uses the ES-1.4.0-and-later compatible
"unmapped_type" sort option.