Extend thrift frame size control to Astyanax
Prior to this commit, there was a way to set the maximum frame size
used in Thrift framed transport (for Cassandra) on the cassandrathrift
backend, but not on the astyanax backend.  The version of astyanax we
used was too old to support this setting.  Upgraded to a newer version
of astyanax that allows setting the frame size and moved the option
into the shared cassandra namespace to reflect its new status as an
option shared between both of the cassandra storage.backend impls.