Use pom.xml ${tinkerpop.version} in asciidoc URLs
The asciidoc source contained URLs linking to various versions of
TinkerPop 3.  It's arguable that this is the better way to go, since
hardcoding a specific TP version into a documentation link means that
the link will never rot, but it could also lead to nonsensical
documentation as Titan continues to track later TP versions throughout
TP's life.  I've left a static image reference with a hardcoded
version, but changed links to prose so that they use
$MAVEN{tinkerpop.version} instead of an arbitrary past TP3 version.