Fix table format in serializer asciidoc
f9e32000a7dbaf17b3bfe88072424aae4a6fad31 seemed to accidentally
introduce a bunch of spaces after commas in the format string for this
page's table, which seems to break it in the HTML output in such a way
that the table appears to stretch off the right side of the page.
Removing the spaces from the formatting string.