Apply C* compaction options only when present
This changes the default behavior of #1156.

Instead of hardcoding SizeTieredCompactionStrategy as the default and
applying it in the absence of an explicit user override, this commit
removes the default value for COMPACTION_STRATEGY.  It also updates
the three use sites to check whether the config contains
COMPACTION_STRATEGY before attempting to read the option's value.

Similarly, if COMPACTION_OPTIONS was not set, or if it was set to an
empty map, then its use sites will not call the associated builder
method, allowing the implementation to use whatever default it might
have (probably empty anyway).