Update versions to sate Spark & Gremlin-Server
This upgrades Jackson from 2.3.0 to 2.4.4 (for compatibility with the
version of Spark preferred by TP 3.1).

This downgrades Jersey from 1.18.2 to 1.9 to match Gremlin-Server.
This is probably going to break something else.

Independent of the version changes, this commit also puts
jersey-client and jersey-guice under version management.  Before, they
were unmanaged, so they could get out of sync with the versions of the
rest of the (managed) jersey artifacts.  Hence jersey-client
mismatched to jersey-server, etc.