Manually commit test transaction
TP 3.1 changed the default tx.close() behavior from commit to
rollback.  This broke TitanGraphBaseTest and the classes that extend
it.  This has been floating in my stash for a while as I swapped
between branches.  It's enough to fix TitanGraphBaseTest and
subclasses, but I am not convinced that it fixes all instances of the
underlying problem (there could be tests that relied on the autocommit
behavior that do not extend TGBT).