Fix test failure caused by #1197
The problem here is that #1197 caused the first job configuration
executed in a JVM to be cached for the life of the JVM.  The tests
execute multiple jobs with distinct configurations while reusing the
same JVM (Hadoop's LocalJobRunner).

This commit changes GiraphInputFormat such that the configuration most
recently supplied to getConf anywhere in the JVM (this is statically
stored) wwill be used any time the refcount of its
TitanVertexDeserializer rises from zero to one.